requirements.txt
pandas
numpy
python-dotenv
matplotlib
sympy
plotly
pybind11
scipy
particle
python-multipart
pyhepmc
fastjet
graphviz
pyyaml